More fixes to the new Gyro
-Gyro resets correctly on teleopInit, autoInit, and robotInit.
(@Competition it should only reset on autoInit)
-PID values were changed. They need to be re-tuned, though. The
competition bot is *much* more powerful than the test bot.
-In cases where the gyro loses connection, the robot currently freaks
out a bit. TODO: Either handle random 0 readings or ignore outliers in
the data. Also, come up with a back up plan if this gyro is borkened,
disconnected, or otherwise unusable.
-After turning suddenly and quickly, the robot overshoots its new
rotation setpoint and corrects itself. This is not gamebreaking, but is
an odd quirk for the drivers to deal with. TODO: Fix this.
-Might need to slow(or ramp in some way) certain actions. This robot is
fast.